Renaming container_builder to agent_runtime Removing poll since it'll become handled in the backend Moved completions under the dev commands Commented some thoughts Commented some streaming return code Tonic might not support atm.
