+++ cli/orb/src/subcommand/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [dependencies]
 structopt = "0.3"
 clap = "2.33"
-container_builder = { path = "../../../../lib/container_builder" }
+agent_runtime = { path = "../../../../lib/agent_runtime" }
 config_parser = { path = "../../../../lib/config_parser" }
 git_meta = { path = "../../../../lib/git_meta" }
 orbital_headers = { path = "../../../../models/orbital_headers" }
