+++ cli/orb/src/subcommand/developer/completion.rs
