+++ cli/orb/src/subcommand/summary.rs
@@ -3,7 +3,7 @@ use structopt::StructOpt;
 
 use crate::{GlobalOption, SubcommandError};
 
-use orbital_headers::build_meta::{client::BuildServiceClient, BuildTarget, BuildSummaryRequest};
+use orbital_headers::build_meta::{client::BuildServiceClient, BuildSummaryRequest, BuildTarget};
 
 use crate::ORB_DEFAULT_URI;
 use tonic::Request;
@@ -28,9 +28,11 @@ pub async fn subcommand_handler(
 
     // Idea: Index should be Option<u32>
     let request = Request::new(BuildSummaryRequest {
-        build: Some(BuildTarget { org : "org_name_goes_here".into(), ..Default::default() }),
+        build: Some(BuildTarget {
+            org: "org_name_goes_here".into(),
+            ..Default::default()
+        }),
         ..Default::default()
-
     });
 
     let response = client.build_summary(request).await?;
