+++ lib/agent_runtime/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "container_builder"
+name = "agent_runtime"
 version = "0.1.0"
 authors = ["T.J. Telan <t.telan@gmail.com>"]
 edition = "2018"
@@ -10,5 +10,5 @@ tokio = "0.1"
 log = "0.4"
 
 [lib]
-name = "container_builder"
+name = "agent_runtime"
 path = "mod.rs"
